Fixing Runtime Issues for Schema.sql
After running the database within a container I realized that the schema
was failing really fast. Had to clean up a bunch of table declarations
in order for the schema to run. Found the following issues:

- A trailing comma in the last field of certain tables
- A typo, where BEGIN turned into BEGINj
- The "geography" column type was not present in the image being used,
so I moved from a simple "postgres" image, into a "postgis" image that
already has the geography plugin installed
- Did not know that constraints need a unique name across the entire
database. This meant changing the name of the primary key for all of the
"many to many" tables, since they all were named "id".

After this I also formatted my schema according to what "sqruff"
believes is best, however, I dont really like the style. Next commit is
probably going to be me trying to get the code to look right with the
formatter